Backend/accounts/serializers.py
from accounts.views import password
from django.db.models import fields
from rest_framework import serializers
from django.contrib.auth import get_user_model

User = get_user_model()

class UserSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ('email', 'password', 'nickname')


class PasswordSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ('password')
        read_only_fields = ('email', 'nickname',)

class ProfileSerializer(serializers.ModelSerializer):
    # 유저의 스크랩 내용
    pass

